move related assets functions to another file
